pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "auto-bml"
version = "0.1.0"
description = "Automated Build-Measure-Learn loops using Google Ads and Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "google-ads>=24.0.0",
    "PyGithub>=2.3.0",
    "click>=8.1.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "PyNaCl>=1.5.0",
]

[project.scripts]
auto-bml = "auto_bml.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_bml*"]
